adjusting type

+++ types.go
@@ -36,15 +36,15 @@ type Weather struct {
 }
 
 type WeatherData struct {
-	Temperature   int    `json:"temperature"`
-	WindDirection string `json:"wind_direction"`
-	WindVelocity  int    `json:"wind_velocity"`
-	Humidity      int    `json:"humidity"`
-	Condition     string `json:"condition"`
-	Pressure      int    `json:"pressure"`
-	Icon          string `json:"icon"`
-	Sensation     int    `json:"sensation"`
-	Date          string `json:"date"`
+	Temperature   int     `json:"temperature"`
+	WindDirection string  `json:"wind_direction"`
+	WindVelocity  float64 `json:"wind_velocity"`
+	Humidity      int     `json:"humidity"`
+	Condition     string  `json:"condition"`
+	Pressure      int     `json:"pressure"`
+	Icon          string  `json:"icon"`
+	Sensation     int     `json:"sensation"`
+	Date          string  `json:"date"`
 }
 
 type ClimaTempoError struct {
